Group styled-components in styles.ts files, fix style issues
