+++ src/views/Ifos/components/Ideology/Ideology.tsx
@@ -1,87 +1,9 @@
 import React from 'react'
-import styled from 'styled-components'
-import { Button, Heading, Text } from '@apeswapfinance/uikit'
+import { Button, Text } from '@apeswapfinance/uikit'
 import useTheme from 'hooks/useTheme'
 import { useHistory } from 'react-router-dom'
 
-const FeatureBox = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-`
-
-const Motto = styled(Text).attrs({
-  fontFamily: 'poppins',
-  textAlign: 'center',
-})`
-  color: ${({ theme }) => theme.colors.yellow};
-  margin: 2px 0 6px 0;
-`
-
-const SectionHeading = styled(Heading)`
-  font-weight: 800;
-`
-
-const Frame = styled.div`
-  margin-top: 48px;
-  margin-bottom: 32px;
-
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-  gap: 20px;
-
-  ${({ theme }) => theme.mediaQueries.sm} {
-    flex-direction: row;
-  }
-`
-
-const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  border-radius: 20px;
-  background-color: ${({ theme }) => theme.colors.card};
-  margin-bottom: 57px;
-  padding: 30px 20px;
-
-  ${({ theme }) => theme.mediaQueries.sm} {
-    padding: 30px 50px;
-  }
-`
-
-const IconBox = styled.div`
-  border-radius: 5px;
-  width: 144px;
-  height: 98px;
-  background-color: ${({ theme }) => (theme.isDark ? '#383838' : '#E5E5E5')};
-  position: relative;
-  flex-shrink: 0;
-  margin-bottom: 8px;
-`
-
-const CenteredImage = styled.img`
-  position: absolute;
-  left: 50%;
-  top: 50%;
-  transform: translate(-50%, -50%);
-`
-const ButtonBox = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-
-  > a,
-  button {
-    width: 260px;
-  }
-
-  ${({ theme }) => theme.mediaQueries.sm} {
-    flex-direction: row;
-    gap: 40px;
-  }
-`
+import { FeatureBox, Motto, SectionHeading, Frame, IconBox, CenteredImage, Container, ButtonBox } from './styles'
 
 interface Props {
   name: string
