+++ src/views/Ifos/components/IfoCard/CardDescription/IfoCardDescription.tsx
@@ -1,6 +1,5 @@
 import React, { useState } from 'react'
-import styled from 'styled-components'
-import { Text, Image, LinkExternal } from '@apeswapfinance/uikit'
+import { StyledIfoCardDescription, Link, ToggleButton, Description, IconImage } from './styles'
 
 export interface IfoCardDescriptionProps {
   defaultIsOpen?: boolean
@@ -8,40 +7,8 @@ export interface IfoCardDescriptionProps {
   projectSiteUrl: string
 }
 
-const StyledIfoCardDescription = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-`
 
-const ToggleButton = styled.button`
-  background: transparent;
-  border: none;
-  color: ${({ theme }) => (theme.isDark ? theme.colors.white : theme.colors.primary)};
-  cursor: pointer;
-  display: block;
-  font-weight: 800;
-  font-family: Poppins;
-  outline: 0;
-  width: 100%;
-`
 
-const Description = styled(Text).attrs({ fontFamily: 'poppins' })<{ isOpen: boolean }>`
-  color: ${({ theme }) => (theme.isDark ? theme.colors.white : theme.colors.primary)};
-  display: ${({ isOpen }) => (isOpen ? 'block' : 'none')};
-  margin-top: 24px;
-`
-
-const IconImage = styled(Image)`
-  margin-left: 8px;
-  display: inline-block;
-`
-
-const Link = styled(LinkExternal)`
-  color: ${({ theme }) => (theme.isDark ? theme.colors.white : theme.colors.primary)};
-`
 
 const IfoCardDescription: React.FC<IfoCardDescriptionProps> = ({
   defaultIsOpen = true,
