+++ src/views/Ifos/components/IfoCard/CardProgress/styles.ts
@@ -1,0 +1,22 @@
+import styled from 'styled-components'
+import { Text } from '@apeswapfinance/uikit'
+
+export const StyledProgress = styled.div`
+  margin-bottom: 16px;
+  margin-top: 15px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  gap: 8px;
+  align-items: center;
+  width: 100%;
+  min-width: 320px;
+`
+
+export const ProgressBar = styled.div`
+  width: 100%;
+`
+
+export const Label = styled(Text)`
+  font-weight: 700;
+`
