+++ src/views/Ifos/components/IfoCard/FourPhaseVesting/IfoCardContribute.tsx
@@ -1,5 +1,4 @@
 import React, { useState } from 'react'
-import styled from 'styled-components'
 import { Button, Text } from '@apeswapfinance/uikit'
 import { useWeb3React } from '@web3-react/core'
 import BigNumber from 'bignumber.js'
@@ -13,8 +12,9 @@ import { ZERO_ADDRESS } from 'config'
 import { CHAIN_ID } from 'config/constants'
 import track from 'utils/track'
 import useUserInfo from './useUserInfo'
+import { VestingButtonWrapper, VestingClaimButton, Claim, DisplayVestingTime, TextWrapRow } from './styles'
 
-import ContributeInput from '../ContributeInput'
+import ContributeInput from '../ContributeInput/ContributeInput'
 
 export interface Props {
   account: string
@@ -27,66 +27,6 @@ export interface Props {
   isActive?: boolean
   isFinished?: boolean
 }
-
-const VestingButtonWrapper = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  justify-content: center;
-  align-items: center;
-`
-
-const VestingClaimButton = styled(Button)`
-  width: 120px;
-  height: 60px;
-  background-color: secondary;
-  margin: 10px;
-  flex-shrink: 0;
-  background: #ffb300;
-  padding: 0;
-  :focus {
-    outline: none !important;
-    box-shadow: none !important;
-    background: #ffb300;
-  }
-  display: flex;
-  flex-direction: column;
-  ${({ theme }) => theme.mediaQueries.lg} {
-    width: 140px;
-  }
-`
-
-const DisplayVestingTime = styled(Text)`
-  font-family: Poppins;
-  font-size: 11px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 26px;
-  ${({ theme }) => theme.mediaQueries.lg} {
-    font-size: 13px;
-  }
-`
-
-const Claim = styled(Text)`
-  font-family: Poppins;
-  font-size: 18px;
-  font-style: normal;
-  font-weight: 700;
-  line-height: 36px;
-  ${({ theme }) => theme.mediaQueries.lg} {
-    font-size: 20px;
-  }
-`
-
-const TextWrapRow = styled.div`
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  flex-direction: row;
-  flex-wrap: wrap;
-`
 
 const formatTime = (time: any): string => {
   return `${time.days}d - ${time.hours}h - ${time.minutes}m`
@@ -192,9 +132,6 @@ const IfoCardContribute: React.FC<Props> = ({
       )}
       {isFinished && (
         <>
-          <Text fontSize="16px" fontFamily="poppins" fontWeight={700} color="textSubtle">
-            You&apos;ll be refunded any excess tokens on your first claim
-          </Text>
           <VestingButtonWrapper>
             {amountContributed > 0 && (
               <>
