+++ src/views/Ifos/components/IfoCard/FourPhaseVesting/index.tsx
@@ -15,9 +15,9 @@ import { getContract } from 'utils/web3'
 import getTimePeriods from 'utils/getTimePeriods'
 import UnlockButton from 'components/UnlockButton'
 import { getBalanceNumber } from 'utils/formatBalance'
-import IfoCardHeader from '../IfoCardHeader'
-import IfoCardProgress from '../IfoCardProgress'
-import IfoCardDetails from '../IfoCardDetails'
+import IfoCardHeader from '../CardHeader/IfoCardHeader'
+import IfoCardProgress from '../CardProgress/IfoCardProgress'
+import IfoCardDetails from '../CardDetails/IfoCardDetails'
 import IfoCardContribute from './IfoCardContribute'
 import useUserInfo from './useUserInfo'
