+++ src/views/Ifos/components/IfoCard/FourPhaseVesting/styles.ts
@@ -1,0 +1,62 @@
+import styled from 'styled-components'
+import { Button, Text } from '@apeswapfinance/uikit'
+
+export const VestingButtonWrapper = styled.div`
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: center;
+  align-items: center;
+`
+
+export const VestingClaimButton = styled(Button)`
+  width: 120px;
+  height: 60px;
+  background-color: secondary;
+  margin: 10px;
+  flex-shrink: 0;
+  background: #ffb300;
+  padding: 0;
+  :focus {
+    outline: none !important;
+    box-shadow: none !important;
+    background: #ffb300;
+  }
+  display: flex;
+  flex-direction: column;
+  ${({ theme }) => theme.mediaQueries.lg} {
+    width: 140px;
+  }
+`
+
+export const DisplayVestingTime = styled(Text)`
+  font-family: Poppins;
+  font-size: 11px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 26px;
+  ${({ theme }) => theme.mediaQueries.lg} {
+    font-size: 13px;
+  }
+`
+
+export const Claim = styled(Text)`
+  font-family: Poppins;
+  font-size: 18px;
+  font-style: normal;
+  font-weight: 700;
+  line-height: 36px;
+  ${({ theme }) => theme.mediaQueries.lg} {
+    font-size: 20px;
+  }
+`
+
+export const TextWrapRow = styled.div`
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  flex-direction: row;
+  flex-wrap: wrap;
+`
