--- src/views/Ifos/components/Title.tsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components'
-import { Heading } from '@apeswapfinance/uikit'
-
-const Title = styled(Heading).attrs({ size: 'lg' })`
-  color: ${({ theme }) => theme.colors.secondary};
-  margin: 16px 0;
-`
-
-export default Title
